Split off simulate and friends to a new properties directory.
Part of #1073
Part of #258
