+++ src/act/properties/b2data.cpp
