+++ src/act/properties/b2data.h
