+++ src/act/properties/min_complex.cpp
