+++ src/act/properties/normalmodes.cpp
@@ -37,13 +37,13 @@
 #include "act/alexandria/alex_modules.h"
 #include "act/alexandria/confighandler.h"
 #include "act/alexandria/molhandler.h"
-#include "act/alexandria/secondvirial.h"
 #include "act/alexandria/train_utility.h"
 #include "act/forcefield/forcefield_xml.h"
 #include "act/import/compound_reader.h"
 #include "act/import/import.h"
 #include "act/molprop/molprop_util.h"
 #include "act/molprop/molprop_xml.h"
+#include "act/properties/secondvirial.h"
 #include "act/utility/jsontree.h"
 #include "act/utility/stringutil.h"
 #include "gromacs/commandline/filenm.h"
