+++ src/act/properties/secondvirial.h
@@ -38,9 +38,9 @@
 #include <vector>
 
 #include "act/alexandria/actmol.h"
-#include "act/alexandria/dimergenerator.h"
 #include "act/forcefield/forcefield.h"
 #include "act/forces/forcecomputer.h"
+#include "act/properties/dimergenerator.h"
 #include "act/utility/jsontree.h"
 #include "gromacs/commandline/pargs.h"
 #include "gromacs/fileio/oenv.h"
