+++ src/act/properties/simulate.cpp
@@ -38,13 +38,13 @@
 #include "act/alexandria/confighandler.h"
 #include "act/alexandria/fragmenthandler.h"
 #include "act/alexandria/molhandler.h"
-#include "act/alexandria/secondvirial.h"
 #include "act/alexandria/train_utility.h"
 #include "act/basics/msg_handler.h"
 #include "act/forcefield/forcefield_xml.h"
 #include "act/import/compound_reader.h"
 #include "act/molprop/molprop_util.h"
 #include "act/molprop/molprop_xml.h"
+#include "act/properties/secondvirial.h"
 #include "act/utility/jsontree.h"
 #include "act/utility/stringutil.h"
 #include "gromacs/commandline/filenm.h"
