+++ src/act/properties/tests/CMakeLists.txt
@@ -32,20 +32,16 @@
 # To help us fund GROMACS development, we humbly ask that you cite
 # the research papers on the package. Check out http://www.gromacs.org.
 
-set(testname "AlexandriaTests")
-set(exename "alexandria-test")
+set(testname "PropertiesTests")
+set(exename "properties-test")
 
 gmx_add_unit_test(
     ${testname}
     ${exename}
     # files with code for tests
-    actmol_util.cpp
-    allbondedsTest.cpp
-    forcecomputerTest.cpp
-    molhandlerTest.cpp
-    openmmTest.cpp
-    topologyTest.cpp
-    # pseudo-library for code for alexandria
+    b2Test.cpp
+    sphereIntegratorTest.cpp
+    # pseudo-library for code for properties
     $<TARGET_OBJECTS:actLib>
     )
 target_link_libraries(${exename} libactgromacs ${GMX_EXTRA_LIBRARIES} ${GMX_EXE_LINKER_FLAGS})
