+++ src/act/properties/velocityhandler.cpp
